총알 갯수에 따른 총알 생성 추가

+++ Assets/Scripts/ScriptableObject/BulletData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class BulletData : ScriptableObject
 {
     [SerializeField]
-    private float count;
-    public float Count { get { return count; } }
+    private int count;
+    public int Count { get { return count; } }
     [SerializeField]
     private bool isPlayer;
     public bool IsPlayer { get { return isPlayer; } }
